+++ src/unittests/floating_point.spec.ts
@@ -2164,6 +2164,187 @@ g.test('smoothStepInterval_f32')
     );
   });
 
+interface ScalarToVectorCase {
+  input: number;
+  expected: (number | IntervalBounds)[];
+}
+
+// Scope for unpack* tests so that they can have constants for magic numbers
+// that don't pollute the global namespace or have unwieldy long names.
+{
+  const kZeroBounds: IntervalBounds = [hexToF32(0x81200000), hexToF32(0x01200000)];
+  const kOneBoundsSnorm: IntervalBounds = [
+    hexToF64(0x3fef_ffff_a000_0000n),
+    hexToF64(0x3ff0_0000_4000_0000n),
+  ];
+  const kOneBoundsUnorm: IntervalBounds = [
+    hexToF64(0x3fef_ffff_b000_0000n),
+    hexToF64(0x3ff0_0000_2800_0000n),
+  ];
+  const kNegOneBoundsSnorm: IntervalBounds = [
+    hexToF64(0xbff0_0000_0000_0000n),
+    hexToF64(0xbfef_ffff_a000_0000n),
+  ];
+
+  const kHalfBounds2x16snorm: IntervalBounds = [
+    hexToF64(0x3fe0_001f_a000_0000n),
+    hexToF64(0x3fe0_0020_8000_0000n),
+  ]; // ~0.5..., due to lack of precision in i16
+  const kNegHalfBounds2x16snorm: IntervalBounds = [
+    hexToF64(0xbfdf_ffc0_6000_0000n),
+    hexToF64(0xbfdf_ffbf_8000_0000n),
+  ]; // ~-0.5..., due to lack of precision in i16
+
+  g.test('unpack2x16snormInterval')
+    .paramsSubcasesOnly<ScalarToVectorCase>(
+      // prettier-ignore
+      [
+        { input: 0x00000000, expected: [kZeroBounds, kZeroBounds] },
+        { input: 0x00007fff, expected: [kOneBoundsSnorm, kZeroBounds] },
+        { input: 0x7fff0000, expected: [kZeroBounds, kOneBoundsSnorm] },
+        { input: 0x7fff7fff, expected: [kOneBoundsSnorm, kOneBoundsSnorm] },
+        { input: 0x80018001, expected: [kNegOneBoundsSnorm, kNegOneBoundsSnorm] },
+        { input: 0x40004000, expected: [kHalfBounds2x16snorm, kHalfBounds2x16snorm] },
+        { input: 0xc001c001, expected: [kNegHalfBounds2x16snorm, kNegHalfBounds2x16snorm] },
+      ]
+    )
+    .fn(t => {
+      const expected = FP.f32.toVector(t.params.expected);
+      const got = FP.f32.unpack2x16snormInterval(t.params.input);
+      t.expect(
+        objectEquals(expected, got),
+        `unpack2x16snormInterval(${t.params.input}) returned [${got}]. Expected [${expected}]`
+      );
+    });
+
+  g.test('unpack2x16floatInterval')
+    .paramsSubcasesOnly<ScalarToVectorCase>(
+      // prettier-ignore
+      [
+        // f16 normals
+        { input: 0x00000000, expected: [0, 0] },
+        { input: 0x80000000, expected: [0, 0] },
+        { input: 0x00008000, expected: [0, 0] },
+        { input: 0x80008000, expected: [0, 0] },
+        { input: 0x00003c00, expected: [1, 0] },
+        { input: 0x3c000000, expected: [0, 1] },
+        { input: 0x3c003c00, expected: [1, 1] },
+        { input: 0xbc00bc00, expected: [-1, -1] },
+        { input: 0x49004900, expected: [10, 10] },
+        { input: 0xc900c900, expected: [-10, -10] },
+
+        // f16 subnormals
+        { input: 0x000003ff, expected: [[0, kValue.f16.subnormal.positive.max], 0] },
+        { input: 0x000083ff, expected: [[kValue.f16.subnormal.negative.min, 0], 0] },
+
+        // f16 out of bounds
+        { input: 0x7c000000, expected: [kAnyBounds, kAnyBounds] },
+        { input: 0xffff0000, expected: [kAnyBounds, kAnyBounds] },
+      ]
+    )
+    .fn(t => {
+      const expected = FP.f32.toVector(t.params.expected);
+      const got = FP.f32.unpack2x16floatInterval(t.params.input);
+      t.expect(
+        objectEquals(expected, got),
+        `unpack2x16floatInterval(${t.params.input}) returned [${got}]. Expected [${expected}]`
+      );
+    });
+
+  const kHalfBounds2x16unorm: IntervalBounds = [
+    hexToF64(0x3fe0_000f_b000_0000n),
+    hexToF64(0x3fe0_0010_7000_0000n),
+  ]; // ~0.5..., due to lack of precision in u16
+
+  g.test('unpack2x16unormInterval')
+    .paramsSubcasesOnly<ScalarToVectorCase>(
+      // prettier-ignore
+      [
+        { input: 0x00000000, expected: [kZeroBounds, kZeroBounds] },
+        { input: 0x0000ffff, expected: [kOneBoundsUnorm, kZeroBounds] },
+        { input: 0xffff0000, expected: [kZeroBounds, kOneBoundsUnorm] },
+        { input: 0xffffffff, expected: [kOneBoundsUnorm, kOneBoundsUnorm] },
+        { input: 0x80008000, expected: [kHalfBounds2x16unorm, kHalfBounds2x16unorm] },
+      ]
+    )
+    .fn(t => {
+      const expected = FP.f32.toVector(t.params.expected);
+      const got = FP.f32.unpack2x16unormInterval(t.params.input);
+      t.expect(
+        objectEquals(expected, got),
+        `unpack2x16unormInterval(${t.params.input}) returned [${got}]. Expected [${expected}]`
+      );
+    });
+
+  const kHalfBounds4x8snorm: IntervalBounds = [
+    hexToF64(0x3fe0_2040_2000_0000n),
+    hexToF64(0x3fe0_2041_0000_0000n),
+  ]; // ~0.50196..., due to lack of precision in i8
+  const kNegHalfBounds4x8snorm: IntervalBounds = [
+    hexToF64(0xbfdf_bf7f_6000_0000n),
+    hexToF64(0xbfdf_bf7e_8000_0000n),
+  ]; // ~-0.49606..., due to lack of precision in i8
+
+  g.test('unpack4x8snormInterval')
+    .paramsSubcasesOnly<ScalarToVectorCase>(
+      // prettier-ignore
+      [
+        { input: 0x00000000, expected: [kZeroBounds, kZeroBounds, kZeroBounds, kZeroBounds] },
+        { input: 0x0000007f, expected: [kOneBoundsSnorm, kZeroBounds, kZeroBounds, kZeroBounds] },
+        { input: 0x00007f00, expected: [kZeroBounds, kOneBoundsSnorm, kZeroBounds, kZeroBounds] },
+        { input: 0x007f0000, expected: [kZeroBounds, kZeroBounds, kOneBoundsSnorm, kZeroBounds] },
+        { input: 0x7f000000, expected: [kZeroBounds, kZeroBounds, kZeroBounds, kOneBoundsSnorm] },
+        { input: 0x00007f7f, expected: [kOneBoundsSnorm, kOneBoundsSnorm, kZeroBounds, kZeroBounds] },
+        { input: 0x7f7f0000, expected: [kZeroBounds, kZeroBounds, kOneBoundsSnorm, kOneBoundsSnorm] },
+        { input: 0x7f007f00, expected: [kZeroBounds, kOneBoundsSnorm, kZeroBounds, kOneBoundsSnorm] },
+        { input: 0x007f007f, expected: [kOneBoundsSnorm, kZeroBounds, kOneBoundsSnorm, kZeroBounds] },
+        { input: 0x7f7f7f7f, expected: [kOneBoundsSnorm, kOneBoundsSnorm, kOneBoundsSnorm, kOneBoundsSnorm] },
+        { input: 0x81818181, expected: [kNegOneBoundsSnorm, kNegOneBoundsSnorm, kNegOneBoundsSnorm, kNegOneBoundsSnorm] },
+        { input: 0x40404040, expected: [kHalfBounds4x8snorm, kHalfBounds4x8snorm, kHalfBounds4x8snorm, kHalfBounds4x8snorm] },
+        { input: 0xc1c1c1c1, expected: [kNegHalfBounds4x8snorm, kNegHalfBounds4x8snorm, kNegHalfBounds4x8snorm, kNegHalfBounds4x8snorm] },
+      ]
+    )
+    .fn(t => {
+      const expected = FP.f32.toVector(t.params.expected);
+      const got = FP.f32.unpack4x8snormInterval(t.params.input);
+      t.expect(
+        objectEquals(expected, got),
+        `unpack4x8snormInterval(${t.params.input}) returned [${got}]. Expected [${expected}]`
+      );
+    });
+
+  const kHalfBounds4x8unorm: IntervalBounds = [
+    hexToF64(0x3fe0_100f_b000_0000n),
+    hexToF64(0x3fe0_1010_7000_0000n),
+  ]; // ~0.50196..., due to lack of precision in u8
+
+  g.test('unpack4x8unormInterval')
+    .paramsSubcasesOnly<ScalarToVectorCase>(
+      // prettier-ignore
+      [
+        { input: 0x00000000, expected: [kZeroBounds, kZeroBounds, kZeroBounds, kZeroBounds] },
+        { input: 0x000000ff, expected: [kOneBoundsUnorm, kZeroBounds, kZeroBounds, kZeroBounds] },
+        { input: 0x0000ff00, expected: [kZeroBounds, kOneBoundsUnorm, kZeroBounds, kZeroBounds] },
+        { input: 0x00ff0000, expected: [kZeroBounds, kZeroBounds, kOneBoundsUnorm, kZeroBounds] },
+        { input: 0xff000000, expected: [kZeroBounds, kZeroBounds, kZeroBounds, kOneBoundsUnorm] },
+        { input: 0x0000ffff, expected: [kOneBoundsUnorm, kOneBoundsUnorm, kZeroBounds, kZeroBounds] },
+        { input: 0xffff0000, expected: [kZeroBounds, kZeroBounds, kOneBoundsUnorm, kOneBoundsUnorm] },
+        { input: 0xff00ff00, expected: [kZeroBounds, kOneBoundsUnorm, kZeroBounds, kOneBoundsUnorm] },
+        { input: 0x00ff00ff, expected: [kOneBoundsUnorm, kZeroBounds, kOneBoundsUnorm, kZeroBounds] },
+        { input: 0xffffffff, expected: [kOneBoundsUnorm, kOneBoundsUnorm, kOneBoundsUnorm, kOneBoundsUnorm] },
+        { input: 0x80808080, expected: [kHalfBounds4x8unorm, kHalfBounds4x8unorm, kHalfBounds4x8unorm, kHalfBounds4x8unorm] },
+      ]
+    )
+    .fn(t => {
+      const expected = FP.f32.toVector(t.params.expected);
+      const got = FP.f32.unpack4x8unormInterval(t.params.input);
+      t.expect(
+        objectEquals(expected, got),
+        `unpack4x8unormInterval(${t.params.input}) returned [${got}]. Expected [${expected}]`
+      );
+    });
+}
+
 interface VectorToIntervalCase {
   input: number[];
   expected: number | IntervalBounds;
