+++ src/webgpu/shader/execution/expression/call/builtin/unpack2x16snorm.spec.ts
@@ -8,10 +8,10 @@ of bits 16×i through 16×i+15 of e as a twos-complement signed integer.
 import { makeTestGroup } from '../../../../../../common/framework/test_group.js';
 import { GPUTest } from '../../../../../gpu_test.js';
 import { TypeF32, TypeU32, TypeVec } from '../../../../../util/conversion.js';
-import { unpack2x16snormInterval } from '../../../../../util/f32_interval.js';
+import { FP } from '../../../../../util/floating_point.js';
 import { fullU32Range } from '../../../../../util/math.js';
 import { makeCaseCache } from '../../case_cache.js';
-import { allInputSources, generateU32ToVectorCases, run } from '../../expression.js';
+import { allInputSources, run } from '../../expression.js';
 
 import { builtin } from './builtin.js';
 
@@ -19,10 +19,18 @@ export const g = makeTestGroup(GPUTest);
 
 export const d = makeCaseCache('unpack2x16snorm', {
   u32_const: () => {
-    return generateU32ToVectorCases(fullU32Range(), 'finite', unpack2x16snormInterval);
+    return FP.f32.generateU32ToIntervalCases(
+      fullU32Range(),
+      'finite',
+      FP.f32.unpack2x16snormInterval
+    );
   },
   u32_non_const: () => {
-    return generateU32ToVectorCases(fullU32Range(), 'unfiltered', unpack2x16snormInterval);
+    return FP.f32.generateU32ToIntervalCases(
+      fullU32Range(),
+      'unfiltered',
+      FP.f32.unpack2x16snormInterval
+    );
   },
 });
